Move interaction interface to the common package
