+++ models/src/main/java/com/vimeo/networking2/AlbumInteractions.kt
@@ -1,5 +1,7 @@
 package com.vimeo.networking2
 
+import com.vimeo.networking2.common.Interaction
+
 /**
  * All actions that can be taken on albums.
  */
