+++ models/src/main/java/com/vimeo/networking2/UserInteractions.kt
@@ -1,6 +1,7 @@
 package com.vimeo.networking2
 
 import com.vimeo.networking2.common.FollowableInteractions
+import com.vimeo.networking2.common.Interaction
 import com.vimeo.networking2.common.UpdatableInteraction
 
 /**
