+++ models/src/main/java/com/vimeo/networking2/common/UpdatableInteraction.kt
@@ -1,6 +1,5 @@
 package com.vimeo.networking2.common
 
-import com.vimeo.networking2.Interaction
 import java.util.*
 
 /**
